split up Kotlin Multiplatform convention plugins, so targets can be applied specifically
