+++ ks3-jdk/build.gradle.kts
@@ -1,27 +1,44 @@
 plugins {
-   id("ks3.conventions.lang.kotlin-jvm")
+   id("ks3.conventions.lang.kotlin-multiplatform-jvm") // Kotlin/JVM only
    id("ks3.conventions.publishing.maven-publish")
 }
 
-dependencies {
-   implementation(platform(libs.kotlin.bom))
-   implementation(platform(libs.kotlinxSerialization.bom))
+kotlin {
+   sourceSets {
+      val commonMain by getting {
+         dependencies {
+            implementation(platform(libs.kotlin.bom))
+            implementation(platform(libs.kotlinxSerialization.bom))
 
-   implementation(projects.ks3Standard)
+            implementation(projects.ks3Standard)
 
-   implementation(libs.kotlinxSerialization.core)
+            implementation(libs.kotlinxSerialization.core)
+         }
+      }
 
-   testImplementation(platform(libs.kotest.bom))
+      val commonTest by getting {
+         dependencies {
+            implementation(platform(libs.kotest.bom))
 
-   testImplementation(kotlin("test"))
+            implementation(kotlin("test"))
 
-   testImplementation(projects.ks3Test)
+            implementation(projects.ks3Test)
 
-   testImplementation(libs.kotest.frameworkEngine)
-   testImplementation(libs.kotest.assertionsCore)
-   testImplementation(libs.kotest.assertionsJson)
-   testImplementation(libs.kotest.property)
+            implementation(libs.kotest.frameworkEngine)
+            implementation(libs.kotest.assertionsCore)
+            implementation(libs.kotest.assertionsJson)
+            implementation(libs.kotest.property)
 
-   testImplementation(libs.kotlinxSerialization.json)
-   testImplementation(libs.kotest.runnerJunit5)
+            implementation(libs.kotlinxSerialization.json)
+         }
+      }
+
+      if (ks3Settings.enableKotlinJvm.get()) {
+         val jvmTest by getting {
+            dependencies {
+               implementation(libs.kotest.runnerJunit5)
+            }
+         }
+      }
+   }
 }
