+++ ks3-test/build.gradle.kts
@@ -1,5 +1,7 @@
 plugins {
-   id("ks3.conventions.lang.kotlin-multiplatform")
+   id("ks3.conventions.lang.kotlin-multiplatform-js")
+   id("ks3.conventions.lang.kotlin-multiplatform-jvm")
+   id("ks3.conventions.lang.kotlin-multiplatform-native")
    id("ks3.conventions.publishing.maven-publish")
 }
